scripts/requirements.txt
# Python dependencies for DynastyRanks scrapers
requests>=2.31.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
